# distsqlrun: optimize merge joiner memory account usage