--- a/pkg/sql/distsqlrun/stream_group_accumulator.go
+++ b/pkg/sql/distsqlrun/stream_group_accumulator.go
@@ -44,7 +44,8 @@ type streamGroupAccumulator struct {
 
 	rowAlloc sqlbase.EncDatumRowAlloc
 
-	memAcc mon.BoundAccount
+	memAcc          mon.BoundAccount
+	minAllocatedSet bool
 }
 
 func makeStreamGroupAccumulator(
@@ -117,7 +118,17 @@ func (s *streamGroupAccumulator) nextGroup(
 			n := len(s.curGroup)
 			ret := s.curGroup[:n:n]
 			s.curGroup = s.curGroup[:0]
-			s.memAcc.Clear(evalCtx.Ctx())
+
+			if !s.minAllocatedSet {
+				err := s.memAcc.SetMinAllocated(evalCtx.Context, int64(row.Size()))
+				if err != nil {
+					return nil, &ProducerMetadata{Err: err}
+				}
+				s.minAllocatedSet = true
+			}
+
+			s.memAcc.Empty(evalCtx.Context)
+
 			s.leftoverRow = row
 			return ret, nil
 		}