--- a/pkg/util/mon/bytes_usage.go
+++ b/pkg/util/mon/bytes_usage.go
@@ -449,7 +449,10 @@ type BoundAccount struct {
 	// reserved is a small buffer to amortize the cost of growing an account. It
 	// decreases as used increases (and vice-versa).
 	reserved int64
-	mon      *BytesMonitor
+	// minAllocated is a minimum allocated bytes size that the account reach
+	// before being able to release bytes.
+	minAllocated int64
+	mon          *BytesMonitor
 }
 
 // MakeStandaloneBudget creates a BoundAccount suitable for root
@@ -468,13 +471,56 @@ func (b BoundAccount) Monitor() *BytesMonitor {
 	return b.mon
 }
 
-func (b BoundAccount) allocated() int64 {
+// Allocated returns the total number of bytes which this account is using or
+// reserving
+func (b BoundAccount) Allocated() int64 {
 	return b.used + b.reserved
 }
 
 // MakeBoundAccount creates a BoundAccount connected to the given monitor.
 func (mm *BytesMonitor) MakeBoundAccount() BoundAccount {
 	return BoundAccount{mon: mm}
+}
+
+// SetMinAllocated allocates a minimum Allocated size (reserved + used) for the
+// account. The account would not be able to Shrink() unless it has surpassed
+// this minimum Allocated value.
+func (b *BoundAccount) SetMinAllocated(ctx context.Context, size int64) error {
+	if size < 0 {
+		panic(fmt.Sprintf("%s: cannot set bound account min allocated to a negative value",
+			b.mon.name))
+	}
+	b.minAllocated = b.mon.roundSize(size)
+
+	if b.Allocated() < b.minAllocated {
+		minExtra := b.mon.roundSize(b.minAllocated - b.Allocated())
+		if err := b.mon.reserveBytes(ctx, minExtra); err != nil {
+			return err
+		}
+		b.reserved += minExtra
+	} else {
+		var released int64
+		if b.used < b.minAllocated {
+			released = b.Allocated() - b.minAllocated
+		} else {
+			released = b.reserved - b.Monitor().poolAllocationSize
+		}
+		b.mon.releaseBytes(ctx, released)
+		b.reserved -= released
+	}
+
+	return nil
+}
+
+// Empty shrinks the account to use 0 bytes, while maintaining the minAllocated
+// size.
+func (b *BoundAccount) Empty(ctx context.Context) {
+	b.reserved += b.used
+	b.used = 0
+	if b.Allocated() > b.minAllocated && b.reserved > b.mon.poolAllocationSize {
+		b.mon.releaseBytes(ctx, b.reserved-b.mon.poolAllocationSize)
+		b.reserved = b.mon.poolAllocationSize
+	}
 }
 
 // Clear releases all the cumulated allocations of an account at once and
@@ -488,6 +534,7 @@ func (b *BoundAccount) Clear(ctx context.Context) {
 	b.Close(ctx)
 	b.used = 0
 	b.reserved = 0
+	b.minAllocated = 0
 }
 
 // Close releases all the cumulated allocations of an account at once.
@@ -497,7 +544,7 @@ func (b *BoundAccount) Close(ctx context.Context) {
 		// monitor -- "bytes out of the aether". This needs not be closed.
 		return
 	}
-	if a := b.allocated(); a > 0 {
+	if a := b.Allocated(); a > 0 {
 		b.mon.releaseBytes(ctx, a)
 	}
 }
@@ -554,7 +601,7 @@ func (b *BoundAccount) Shrink(ctx context.Context, delta int64) {
 	}
 	b.used -= delta
 	b.reserved += delta
-	if b.reserved >= b.mon.poolAllocationSize {
+	if b.Allocated() > b.minAllocated && b.reserved > b.mon.poolAllocationSize {
 		b.mon.releaseBytes(ctx, b.reserved-b.mon.poolAllocationSize)
 		b.reserved = b.mon.poolAllocationSize
 	}
@@ -611,6 +658,10 @@ func (mm *BytesMonitor) reserveBytes(ctx context.Context, x int64) error {
 // releaseBytes releases bytes previously successfully registered via
 // reserveBytes().
 func (mm *BytesMonitor) releaseBytes(ctx context.Context, sz int64) {
+	if sz == 0 {
+		return
+	}
+
 	mm.mu.Lock()
 	defer mm.mu.Unlock()
 	if mm.mu.curAllocated < sz {
@@ -665,7 +716,7 @@ func (mm *BytesMonitor) roundSize(sz int64) int64 {
 func (mm *BytesMonitor) releaseBudget(ctx context.Context) {
 	// NB: mm.mu need not be locked here, as this is only called from StopMonitor().
 	if log.V(2) {
-		log.Infof(ctx, "%s: releasing %d bytes to the pool", mm.name, mm.mu.curBudget.allocated())
+		log.Infof(ctx, "%s: releasing %d bytes to the pool", mm.name, mm.mu.curBudget.Allocated())
 	}
 	mm.mu.curBudget.Clear(ctx)
 }