Fixes #30687.

Instead of opening and closing the memory account on every single row of the merge join, we can allocate a certain minimum block size on the memory account. This way, we are not constantly requesting and releasing memory.

Benchmark results for `BenchmarkMergeJoiner`:

```
name                           old time/op    new time/op    delta
MergeJoiner/InputSize=0-8        5.37µs ±24%    5.85µs ±32%      ~     (p=0.280 n=10+10)
MergeJoiner/InputSize=4-8        14.7µs ±27%     8.4µs ± 7%   -42.88%  (p=0.000 n=10+9)
MergeJoiner/InputSize=16-8       22.9µs ±20%    12.7µs ±24%   -44.44%  (p=0.000 n=10+9)
MergeJoiner/InputSize=256-8       210µs ± 7%      89µs ± 4%   -57.62%  (p=0.000 n=10+9)
MergeJoiner/InputSize=4096-8     3.84ms ±17%    1.31ms ± 4%   -65.80%  (p=0.000 n=10+9)
MergeJoiner/InputSize=65536-8    54.3ms ±12%    22.5ms ± 6%   -58.51%  (p=0.000 n=10+10)

name                           old alloc/op   new alloc/op   delta
MergeJoiner/InputSize=0-8        6.65kB ± 0%    6.66kB ± 0%    +0.24%  (p=0.000 n=10+10)
MergeJoiner/InputSize=4-8        9.72kB ± 0%    9.74kB ± 0%    +0.16%  (p=0.000 n=10+10)
MergeJoiner/InputSize=16-8       9.72kB ± 0%    9.74kB ± 0%    +0.16%  (p=0.000 n=10+10)
MergeJoiner/InputSize=256-8      32.8kB ± 0%    32.8kB ± 0%    +0.05%  (p=0.000 n=10+10)
MergeJoiner/InputSize=4096-8      401kB ± 0%     401kB ± 0%    +0.00%  (p=0.000 n=10+10)
MergeJoiner/InputSize=65536-8    6.30MB ± 0%    6.30MB ± 0%    +0.00%  (p=0.000 n=9+10)

name                           old allocs/op  new allocs/op  delta
MergeJoiner/InputSize=0-8          15.0 ± 0%      15.0 ± 0%      ~     (all equal)
MergeJoiner/InputSize=4-8          17.0 ± 0%      17.0 ± 0%      ~     (all equal)
MergeJoiner/InputSize=16-8         17.0 ± 0%      17.0 ± 0%      ~     (all equal)
MergeJoiner/InputSize=256-8        47.0 ± 0%      47.0 ± 0%      ~     (all equal)
MergeJoiner/InputSize=4096-8        527 ± 0%       527 ± 0%      ~     (all equal)
MergeJoiner/InputSize=65536-8     8.21k ± 0%     8.21k ± 0%      ~     (all equal)

name                           old speed      new speed      delta
MergeJoiner/InputSize=4-8      4.42MB/s ±22%  7.65MB/s ± 7%   +72.86%  (p=0.000 n=10+9)
MergeJoiner/InputSize=16-8     11.4MB/s ±23%  20.3MB/s ±20%   +77.91%  (p=0.000 n=10+9)
MergeJoiner/InputSize=256-8    19.5MB/s ± 7%  46.0MB/s ± 4%  +135.56%  (p=0.000 n=10+9)
MergeJoiner/InputSize=4096-8   17.3MB/s ±19%  49.9MB/s ± 4%  +187.63%  (p=0.000 n=10+9)
MergeJoiner/InputSize=65536-8  19.4MB/s ±11%  46.6MB/s ± 6%  +140.58%  (p=0.000 n=10+10)
```