@jordanlewis 2018-10-03
> Instead resizing the to a 0, we can achieve significant improvement in speed if we if we reserve a minimum size for the MergeJoiner. ie. we grow the account to merge_joiner_min_size at the start, and then s.memAcc.ResizeTo(evalCtx.Context, int64(merge_joiner_min_size)) here. Thoughts?

That does seem good, but why is that such an improvement?