@changangela 2018-10-10
Thanks @knz for the review! I am working on the changes and will create another PR.